dobby_gRPC/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.10
importlib_metadata==8.5.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.7
MarkupSafe==3.0.2
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
setuptools==58.1.0
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.3
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==2.2.3
Werkzeug==3.1.2
wheel==0.44.0
wrapt==1.12.1
zipp==3.20.2
